risk-analysis-unit/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pandas==2.1.3
scikit-learn==1.5.0
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu
requests==2.32.4
apscheduler==3.10.4
python-dotenv==1.0.0
numpy>=1.24.0
Pillow>=10.0.0
httpx>=0.25.0
